GO3. Bankstore. Added some API.

+++ GO3/bankstore/db/migrations/000001_create_table.up.sql
@@ -3,7 +3,7 @@ CREATE TYPE Currency AS ENUM ('USD', 'EUR', 'RUB');
 CREATE TABLE IF NOT EXISTS accounts (
     id BIGSERIAL PRIMARY KEY,
     owner VARCHAR(255) NOT NULL,
-    balance DECIMAL(10, 2) NOT NULL,
+    balance BIGINT NOT NULL,
     currency Currency NOT NULL,
     created TIMESTAMP NOT NULL default now()
 );
@@ -11,7 +11,7 @@ CREATE TABLE IF NOT EXISTS accounts (
 CREATE TABLE entries (
     id BIGSERIAL PRIMARY KEY,
     account_id BIGINT NOT NULL REFERENCES accounts(id) ON DELETE CASCADE,
-    amount DECIMAL(10, 2) NOT NULL,
+    amount BIGINT NOT NULL,
     created TIMESTAMP NOT NULL default now()
 );
 
@@ -19,7 +19,7 @@ CREATE TABLE transfers (
     id BIGSERIAL PRIMARY KEY,
     from_account_id BIGINT NOT NULL REFERENCES accounts(id) ON DELETE CASCADE,
     to_account_id BIGINT NOT NULL REFERENCES accounts(id) ON DELETE CASCADE,
-    amount DECIMAL(10, 2) NOT NULL,
+    amount BIGINT NOT NULL,
     created TIMESTAMP NOT NULL default now()
 );
